+++ packages/firestore/memory/package.json
@@ -1,11 +1,11 @@
 {
   "name": "@firebase/firestore/memory",
   "description": "A memory-only build of the Cloud Firestore JS SDK.",
-  "main": "../dist/index.memory.node.cjs.js",
-  "main-esm2017": "../dist/index.memory.node.esm2017.js",
+  "main": "../dist/node-cjs/memory.js",
+  "main-esm2017": "../dist/node-esm2017/memory.js",
   "react-native": "../dist/index.memory.rn.esm2017.js",
-  "browser": "../dist/index.memory.esm.js",
-  "module": "../dist/index.memory.esm.js",
-  "esm2017": "../dist/index.memory.esm2017.js",
+  "browser": "../dist/esm5/memory.js",
+  "module": "../dist/esm5/memory.js",
+  "esm2017": "../dist/esm2017/memory.js",
   "typings": "../dist/index.d.ts"
 }
