+++ packages/firestore/package.json
@@ -17,6 +17,8 @@
     "build:lite": "rollup -c rollup.config.lite.js",
     "build:bundle": "rollup -c rollup.config.bundle.js",
     "build:exp:release": "yarn build:exp && yarn build:lite",
+    "build:rn": "rollup -c rollup.config.rn.js",
+    "build:browser": "rollup -c rollup.config.browser.js",
     "predev": "yarn prebuild",
     "dev": "rollup -c -w",
     "lint": "eslint -c .eslintrc.js '**/*.ts' --ignore-path '../../.gitignore'",
@@ -52,14 +54,12 @@
     "predoc": "node ../../scripts/exp/remove-exp.js temp",
     "doc": "api-documenter markdown --input temp --output docs"
   },
-  "main": "dist/index.node.cjs.js",
-  "main-esm2017": "dist/index.node.esm2017.js",
+  "main": "dist/node-cjs/index.js",
+  "main-esm2017": "dist/node-esm2017/index.js",
   "react-native": "dist/index.rn.esm2017.js",
-  "browser": "dist/index.esm.js",
-  "module": "dist/index.esm.js",
-  "esm2017": "dist/index.esm2017.js",
-  "bundle": "dist/index.bundle.js",
-  "bundleModule": "dist/index.bundle.module.js",
+  "browser": "dist/esm5/index.js",
+  "module": "dist/esm5/index.js",
+  "esm2017": "dist/esm2017/index.js",
   "license": "Apache-2.0",
   "files": [
     "dist",
@@ -87,7 +87,7 @@
     "json-stable-stringify": "1.0.1",
     "protobufjs": "6.10.1",
     "rollup": "2.33.1",
-    "rollup-plugin-copy-assets": "1.1.0",
+    "rollup-plugin-copy-assets": "2.0.3",
     "@rollup/plugin-json": "4.1.0",
     "@rollup/plugin-node-resolve": "9.0.0",
     "rollup-plugin-replace": "2.2.0",
