+++ packages/firestore/rollup.config.browser.js
@@ -16,14 +16,21 @@
  */
 
 import pkg from './package.json';
+import bundlePkg from './bundle/package.json';
+import memoryPkg from './memory/package.json';
+import path from 'path';
 
 const util = require('./rollup.shared');
 
 export default [
   {
-    input: 'index.ts',
+    input: {
+      index: 'index.ts',
+      memory: 'index.memory.ts',
+      bundle: 'index.bundle.ts'
+    },
     output: {
-      file: pkg.esm2017,
+      dir: 'dist/esm2017',
       format: 'es',
       sourcemap: true
     },
@@ -34,8 +41,16 @@ export default [
     }
   },
   {
-    input: pkg.esm2017,
-    output: { file: pkg.module, format: 'es', sourcemap: true },
+    input: {
+      index: pkg.esm2017,
+      memory: path.resolve('./memory', memoryPkg.esm2017),
+      bundle: path.resolve('./bundle', bundlePkg.esm2017)
+    },
+    output: {
+      dir: 'dist/esm5',
+      format: 'es',
+      sourcemap: true
+    },
     plugins: util.es2017ToEs5Plugins(/* mangled= */ true),
     external: util.resolveBrowserExterns,
     treeshake: {
