+++ packages/firestore/src/core/bundle.ts
@@ -29,18 +29,37 @@ import { BundleMetadata } from '../protos/firestore_bundle_proto';
 import * as api from '../protos/firestore_proto_api';
 import { DocumentKey } from '../model/document_key';
 import { MaybeDocument, NoDocument } from '../model/document';
-import { debugAssert } from '../util/assert';
-import {
-  applyBundleDocuments,
-  LocalStore,
-  saveNamedQuery
-} from '../local/local_store';
+import { debugAssert, debugCast } from '../util/assert';
+import { LocalStore } from '../local/local_store';
 import { SizedBundleElement } from '../util/bundle_reader';
 import {
   documentKeySet,
   DocumentKeySet,
   MaybeDocumentMap
 } from '../model/collections';
+import {
+  FirestoreClient,
+  getLocalStore,
+  getSyncEngine
+} from './firestore_client';
+import { LoadBundleTask } from '../api/bundle';
+import { BundleReader } from '../util/bundle_reader';
+import { newSerializer, newTextEncoder } from '../platform/serializer';
+import { toByteStreamReader } from '../platform/byte_stream_reader';
+import {
+  emitNewSnapsAndNotifyLocalStore,
+  SyncEngine,
+  SyncEngineImpl
+} from './sync_engine';
+import { logWarn } from '../util/log';
+import { LOG_TAG } from '../../lite/src/api/components';
+import {
+  applyBundleDocuments,
+  getNamedQuery,
+  hasNewerBundle,
+  saveBundle,
+  saveNamedQuery
+} from '../local/local_store_bundle';
 
 /**
  * Represents a Firestore bundle saved by the SDK in its local storage.
@@ -116,7 +135,7 @@ export class BundleConverter {
  * Returns a `LoadBundleTaskProgress` representing the initial progress of
  * loading a bundle.
  */
-export function bundleInitialProgress(
+function bundleInitialProgress(
   metadata: BundleMetadata
 ): firestore.LoadBundleTaskProgress {
   return {
@@ -132,7 +151,7 @@ export function bundleInitialProgress(
  * Returns a `LoadBundleTaskProgress` representing the progress that the loading
  * has succeeded.
  */
-export function bundleSuccessProgress(
+function bundleSuccessProgress(
   metadata: BundleMetadata
 ): firestore.LoadBundleTaskProgress {
   return {
@@ -144,7 +163,7 @@ export function bundleSuccessProgress(
   };
 }
 
-export class BundleLoadResult {
+class BundleLoadResult {
   constructor(
     readonly progress: firestore.LoadBundleTaskProgress,
     readonly changedDocs: MaybeDocumentMap
@@ -155,7 +174,7 @@ export class BundleLoadResult {
  * A class to process the elements from a bundle, load them into local
  * storage and provide progress update while loading.
  */
-export class BundleLoader {
+class BundleLoader {
   /** The current progress of loading */
   private progress: firestore.LoadBundleTaskProgress;
   /** Batched queries to be saved into storage */
@@ -260,5 +279,115 @@ export class BundleLoader {
 
     this.progress.taskState = 'Success';
     return new BundleLoadResult({ ...this.progress }, changedDocuments);
+  }
+}
+
+export async function firestoreClientLoadBundle(
+  client: FirestoreClient,
+  data: ReadableStream<Uint8Array> | ArrayBuffer | string,
+  resultTask: LoadBundleTask
+): Promise<void> {
+  const reader = createBundleReader(
+    data,
+    newSerializer((await client.getConfiguration()).databaseInfo.databaseId)
+  );
+  client.asyncQueue.enqueueAndForget(async () => {
+    syncEngineLoadBundle(await getSyncEngine(client), reader, resultTask);
+  });
+}
+
+export function firestoreClientGetNamedQuery(
+  client: FirestoreClient,
+  queryName: string
+): Promise<NamedQuery | undefined> {
+  return client.asyncQueue.enqueue(async () =>
+    getNamedQuery(await getLocalStore(client), queryName)
+  );
+}
+
+function createBundleReader(
+  data: ReadableStream<Uint8Array> | ArrayBuffer | string,
+  serializer: JsonProtoSerializer
+): BundleReader {
+  let content: ReadableStream<Uint8Array> | ArrayBuffer;
+  if (typeof data === 'string') {
+    content = newTextEncoder().encode(data);
+  } else {
+    content = data;
+  }
+  return new BundleReader(toByteStreamReader(content), serializer);
+}
+
+/**
+ * Loads a Firestore bundle into the SDK. The returned promise resolves when
+ * the bundle finished loading.
+ *
+ * @param bundleReader Bundle to load into the SDK.
+ * @param task LoadBundleTask used to update the loading progress to public API.
+ */
+function syncEngineLoadBundle(
+  syncEngine: SyncEngine,
+  bundleReader: BundleReader,
+  task: LoadBundleTask
+): void {
+  const syncEngineImpl = debugCast(syncEngine, SyncEngineImpl);
+
+  // eslint-disable-next-line @typescript-eslint/no-floating-promises
+  loadBundleImpl(syncEngineImpl, bundleReader, task).then(() => {
+    syncEngineImpl.sharedClientState.notifyBundleLoaded();
+  });
+}
+
+async function loadBundleImpl(
+  syncEngine: SyncEngineImpl,
+  reader: BundleReader,
+  task: LoadBundleTask
+): Promise<void> {
+  try {
+    const metadata = await reader.getMetadata();
+    const skip = await hasNewerBundle(syncEngine.localStore, metadata);
+    if (skip) {
+      await reader.close();
+      task._completeWith(bundleSuccessProgress(metadata));
+      return;
+    }
+
+    task._updateProgress(bundleInitialProgress(metadata));
+
+    const loader = new BundleLoader(
+      metadata,
+      syncEngine.localStore,
+      reader.serializer
+    );
+    let element = await reader.nextElement();
+    while (element) {
+      debugAssert(
+        !element.payload.metadata,
+        'Unexpected BundleMetadata element.'
+      );
+      const progress = await loader.addSizedElement(element);
+      if (progress) {
+        task._updateProgress(progress);
+      }
+
+      element = await reader.nextElement();
+    }
+
+    const result = await loader.complete();
+    // TODO(b/160876443): This currently raises snapshots with
+    // `fromCache=false` if users already listen to some queries and bundles
+    // has newer version.
+    await emitNewSnapsAndNotifyLocalStore(
+      syncEngine,
+      result.changedDocs,
+      /* remoteEvent */ undefined
+    );
+
+    // Save metadata, so loading the same bundle will skip.
+    await saveBundle(syncEngine.localStore, metadata);
+    task._completeWith(result.progress);
+  } catch (e) {
+    logWarn(LOG_TAG, `Loading bundle failed with ${e}`);
+    task._failWith(e);
   }
 }
